+++ example/data_source_virtual_environment_cluster_aliases.tf
@@ -2,8 +2,6 @@ data "proxmox_virtual_environment_cluster_aliases" "example" {
   depends_on = [proxmox_virtual_environment_cluster_alias.example]
 }
 
-output "data_proxmox_virtual_environment_cluster_aliases" {
-  value = {
-    "alias_ids" = data.proxmox_virtual_environment_cluster_aliases.example.alias_ids
-  }
+output "data_proxmox_virtual_environment_cluster_aliases_example" {
+  value = data.proxmox_virtual_environment_cluster_aliases.example
 }