+++ example/data_source_virtual_environment_time.tf
@@ -2,10 +2,6 @@ data "proxmox_virtual_environment_time" "example" {
   node_name = data.proxmox_virtual_environment_nodes.example.names[0]
 }
 
-output "data_proxmox_virtual_environment_time" {
-  value = {
-    local_time = data.proxmox_virtual_environment_time.example.local_time
-    time_zone  = data.proxmox_virtual_environment_time.example.time_zone
-    utc_time   = data.proxmox_virtual_environment_time.example.utc_time
-  }
+output "data_proxmox_virtual_environment_time_example" {
+  value = data.proxmox_virtual_environment_time.example
 }