+++ example/resource_virtual_environment_pool.tf
@@ -3,14 +3,6 @@ resource "proxmox_virtual_environment_pool" "example" {
   pool_id = "terraform-provider-proxmox-example"
 }
 
-output "resource_proxmox_virtual_environment_pool_example_comment" {
-  value = proxmox_virtual_environment_pool.example.comment
-}
-
-output "resource_proxmox_virtual_environment_pool_example_members" {
-  value = proxmox_virtual_environment_pool.example.members
-}
-
-output "resource_proxmox_virtual_environment_pool_example_pool_id" {
-  value = proxmox_virtual_environment_pool.example.id
+output "resource_proxmox_virtual_environment_pool_example" {
+  value = proxmox_virtual_environment_pool.example
 }