+++ example/resource_virtual_environment_role.tf
@@ -5,10 +5,6 @@ resource "proxmox_virtual_environment_role" "example" {
   role_id = "terraform-provider-proxmox-example"
 }
 
-output "resource_proxmox_virtual_environment_role_example_privileges" {
-  value = proxmox_virtual_environment_role.example.privileges
-}
-
-output "resource_proxmox_virtual_environment_role_example_role_id" {
-  value = proxmox_virtual_environment_role.example.role_id
+output "resource_proxmox_virtual_environment_role_example" {
+  value = proxmox_virtual_environment_role.example
 }