+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 ## v0.5.0
 
+FEATURES:
+
+* **New Resource:** `proxmox_virtual_environment_user_token`
+
 ENHANCEMENTS:
 
+* data-source/user: Add `tokens` attribute
+* data-source/users: Add `tokens` attribute
 * provider/configuration: Add `virtual_environment.token` argument for API token support
+* resource/user: Add `tokens` attribute
+
+BUG FIXES:
+
+* data-source/users: Fix empty `groups` attribute
 
 ## v0.4.2